Refactor SSH mixins and update modules
