+++ lib/msf/core/exploit/mixins.rb
@@ -121,8 +121,5 @@ require 'msf/core/exploit/http/jboss'
 # Kerberos Support
 require 'msf/core/exploit/kerberos/client'
 
-# Fortinet
-require 'msf/core/exploit/fortinet'
-
 # Other
 require 'msf/core/exploit/windows_constants'
