+++ lib/msf/core/exploit/ssh.rb
@@ -1,8 +1,18 @@
-module Msf
-  module Exploit::Remote::SSH
-    require 'rex/socket/ssh_factory'
-    def ssh_socket_factory
-      Rex::Socket::SSHFactory.new(framework, self, datastore['Proxies'])
-    end
+module Msf::Exploit::Remote::SSH
+
+  # Require most things so that modules using this will "just work"
+  require 'net/ssh'
+  require 'net/ssh/command_stream'
+  require 'rex/socket/ssh_factory'
+  require 'msf/core/exploit/ssh/auth_methods'
+
+  def ssh_socket_factory
+    Rex::Socket::SSHFactory.new(framework, self, datastore['Proxies'])
   end
+
+  # Finally patch in our custom auth methods:
+  #   malformed-packet
+  #   fortinet-backdoor
+  include Msf::Exploit::Remote::SSH::AuthMethods
+
 end
